Rename project from jaxb-java-time-adapters to threeten-jaxb Renames package com.migesok.jaxb.adapter.javatime to org.threetenjaxb.threetenjaxb
