+++ README.md
@@ -1,7 +1,7 @@
 JAXB adapters for Java 8 Date and Time API (JSR-310) types
 ---------
 
-[![Build Status](https://travis-ci.com/jaxb-java-time-adapters/jaxb-java-time-adapters.svg?branch=master)](https://travis-ci.com/jaxb-java-time-adapters/jaxb-java-time-adapters)
+[![Build Status](https://travis-ci.com/threeten-jaxb/threeten-jaxb.svg?branch=master)](https://travis-ci.com/threeten-jaxb/threeten-jaxb)
 
 New Date and Time API introduced in Java 8 (aka JSR-310) is a big step forward comparing to what were available before in the JVM standard library. Sad thing that Java Architecture for XML Binding (JAXB) by default do not speak the new types. This library aimes to address the issue by providing a collection of [type adapters](http://docs.oracle.com/javase/8/docs/api/javax/xml/bind/annotation/adapters/XmlAdapter.html) to enable use of Java 8 Date and Time API types in JAXB bindings.
 
@@ -38,7 +38,7 @@ Source code is licensed under Apache 2.0 licence.
 Release 1.1.3 is the current latest release.
 It depends on JDK 1.8 or later.
 
-Available in the [Maven Central repository](http://search.maven.org/#artifactdetails|com.migesok|jaxb-java-time-adapters|1.1.3|jar)
+Available in the [Maven Central repository](https://search.maven.org/artifact/com.migesok/jaxb-java-time-adapters/1.1.3/jar)
 
 ### Support
 Please use GitHub issues and Pull Requests for support.
