--- src/main/java/com/migesok/jaxb/adapter/javatime/LocalDateTimeXmlAdapter.java
@@ -1,18 +1,0 @@
-package com.migesok.jaxb.adapter.javatime;
-
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-
-/**
- * {@code XmlAdapter} mapping JSR-310 {@code LocalDateTime} to ISO-8601 string
- * <p>
- * String format details: {@link java.time.format.DateTimeFormatter#ISO_LOCAL_DATE_TIME}
- *
- * @see javax.xml.bind.annotation.adapters.XmlAdapter
- * @see java.time.LocalDateTime
- */
-public class LocalDateTimeXmlAdapter extends TemporalAccessorXmlAdapter<LocalDateTime> {
-    public LocalDateTimeXmlAdapter() {
-        super(DateTimeFormatter.ISO_LOCAL_DATE_TIME, LocalDateTime::from);
-    }
-}
