+++ src/main/java/org/threetenjaxb/threetenjaxb/LocalDateXmlAdapter.java
@@ -1,0 +1,19 @@
+package org.threetenjaxb.threetenjaxb;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+/**
+ * {@code XmlAdapter} mapping JSR-310 {@code LocalDate} to ISO-8601 string
+ * <p>
+ * It uses {@link java.time.format.DateTimeFormatter#ISO_DATE} for parsing and serializing,
+ * time-zone information ignored.
+ *
+ * @see javax.xml.bind.annotation.adapters.XmlAdapter
+ * @see java.time.LocalDate
+ */
+public class LocalDateXmlAdapter extends TemporalAccessorXmlAdapter<LocalDate> {
+    public LocalDateXmlAdapter() {
+        super(DateTimeFormatter.ISO_DATE, LocalDate::from);
+    }
+}
