--- src/test/java/com/migesok/jaxb/adapter/javatime/DurationXmlAdapterTest.java
@@ -1,19 +1,0 @@
-package com.migesok.jaxb.adapter.javatime;
-
-import java.time.Duration;
-
-public class DurationXmlAdapterTest extends AbstractToStringAdapterTest<Duration, DurationXmlAdapter> {
-    public DurationXmlAdapterTest() {
-        super(DurationXmlAdapter.class);
-    }
-
-    @Override
-    protected Duration getNotNullValue() {
-        return Duration.parse("P2DT3H4M");
-    }
-
-    @Override
-    protected String getNotValidStringValue() {
-        return "blah-blah";
-    }
-}
