--- src/test/java/com/migesok/jaxb/adapter/javatime/LocalDateXmlAdapterTest.java
@@ -1,19 +1,0 @@
-package com.migesok.jaxb.adapter.javatime;
-
-import java.time.LocalDate;
-
-public class LocalDateXmlAdapterTest extends AbstractToStringAdapterTest<LocalDate, LocalDateXmlAdapter> {
-    public LocalDateXmlAdapterTest() {
-        super(LocalDateXmlAdapter.class);
-    }
-
-    @Override
-    protected LocalDate getNotNullValue() {
-        return LocalDate.of(2014, 12, 31);
-    }
-
-    @Override
-    protected String getNotValidStringValue() {
-        return "blah-blah";
-    }
-}
