--- src/test/java/com/migesok/jaxb/adapter/javatime/YearMonthXmlAdapterTest.java
@@ -1,19 +1,0 @@
-package com.migesok.jaxb.adapter.javatime;
-
-import java.time.YearMonth;
-
-public class YearMonthXmlAdapterTest extends AbstractToStringAdapterTest<YearMonth, YearMonthXmlAdapter> {
-    public YearMonthXmlAdapterTest() {
-        super(YearMonthXmlAdapter.class);
-    }
-
-    @Override
-    protected YearMonth getNotNullValue() {
-        return YearMonth.of(2014, 12);
-    }
-
-    @Override
-    protected String getNotValidStringValue() {
-        return "blah-blah";
-    }
-}
