+++ src/test/java/org/threetenjaxb/threetenjaxb/LocalDateXmlAdapterTest.java
@@ -1,0 +1,19 @@
+package org.threetenjaxb.threetenjaxb;
+
+import java.time.LocalDate;
+
+public class LocalDateXmlAdapterTest extends AbstractToStringAdapterTest<LocalDate, LocalDateXmlAdapter> {
+    public LocalDateXmlAdapterTest() {
+        super(LocalDateXmlAdapter.class);
+    }
+
+    @Override
+    protected LocalDate getNotNullValue() {
+        return LocalDate.of(2014, 12, 31);
+    }
+
+    @Override
+    protected String getNotValidStringValue() {
+        return "blah-blah";
+    }
+}
