+++ src/test/java/org/threetenjaxb/threetenjaxb/PeriodXmlAdapterTest.java
@@ -1,0 +1,19 @@
+package org.threetenjaxb.threetenjaxb;
+
+import java.time.Period;
+
+public class PeriodXmlAdapterTest extends AbstractToStringAdapterTest<Period, PeriodXmlAdapter> {
+    public PeriodXmlAdapterTest() {
+        super(PeriodXmlAdapter.class);
+    }
+
+    @Override
+    protected Period getNotNullValue() {
+        return Period.parse("P1Y2M3W4D");
+    }
+
+    @Override
+    protected String getNotValidStringValue() {
+        return "blah-blah";
+    }
+}
