+++ src/test/java/org/threetenjaxb/threetenjaxb/YearMonthXmlAdapterTest.java
@@ -1,0 +1,19 @@
+package org.threetenjaxb.threetenjaxb;
+
+import java.time.YearMonth;
+
+public class YearMonthXmlAdapterTest extends AbstractToStringAdapterTest<YearMonth, YearMonthXmlAdapter> {
+    public YearMonthXmlAdapterTest() {
+        super(YearMonthXmlAdapter.class);
+    }
+
+    @Override
+    protected YearMonth getNotNullValue() {
+        return YearMonth.of(2014, 12);
+    }
+
+    @Override
+    protected String getNotValidStringValue() {
+        return "blah-blah";
+    }
+}
